Rename PlatformPlatform.SharedKernel.Entities to PlatformPlatform.SharedKernel.Domain
