+++ application/account-management/Core/Authentication/Commands/Logout.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Http;
 using PlatformPlatform.AccountManagement.Authentication.Services;
 using PlatformPlatform.AccountManagement.TelemetryEvents;
 using PlatformPlatform.SharedKernel.Cqrs;
-using PlatformPlatform.SharedKernel.Entities;
+using PlatformPlatform.SharedKernel.Domain;
 using PlatformPlatform.SharedKernel.TelemetryEvents;
 
 namespace PlatformPlatform.AccountManagement.Authentication.Commands;
