+++ application/account-management/Core/Authentication/Domain/Login.cs
@@ -1,6 +1,6 @@
 using JetBrains.Annotations;
 using PlatformPlatform.AccountManagement.Users.Domain;
-using PlatformPlatform.SharedKernel.Entities;
+using PlatformPlatform.SharedKernel.Domain;
 using PlatformPlatform.SharedKernel.IdGenerators;
 
 namespace PlatformPlatform.AccountManagement.Authentication.Domain;
