+++ application/account-management/Core/Signups/Domain/Signup.cs
@@ -1,5 +1,5 @@
 using JetBrains.Annotations;
-using PlatformPlatform.SharedKernel.Entities;
+using PlatformPlatform.SharedKernel.Domain;
 using PlatformPlatform.SharedKernel.IdGenerators;
 
 namespace PlatformPlatform.AccountManagement.Signups.Domain;
