+++ application/account-management/Core/Tenants/Domain/Tenant.cs
@@ -1,4 +1,4 @@
-using PlatformPlatform.SharedKernel.Entities;
+using PlatformPlatform.SharedKernel.Domain;
 
 namespace PlatformPlatform.AccountManagement.Tenants.Domain;
