+++ application/account-management/Core/Tenants/Domain/TenantEvents.cs
@@ -1,5 +1,5 @@
+using PlatformPlatform.SharedKernel.Domain;
 using PlatformPlatform.SharedKernel.DomainEvents;
-using PlatformPlatform.SharedKernel.Entities;
 
 namespace PlatformPlatform.AccountManagement.Tenants.Domain;
