+++ application/shared-kernel/SharedKernel/Domain/EntityEqualityComparer.cs
@@ -1,4 +1,4 @@
-namespace PlatformPlatform.SharedKernel.Entities;
+namespace PlatformPlatform.SharedKernel.Domain;
 
 public sealed class EntityEqualityComparer<T> : IEqualityComparer<Entity<T>> where T : IComparable<T>
 {
