+++ application/shared-kernel/SharedKernel/DomainEvents/IDomainEventCollector.cs
@@ -1,4 +1,4 @@
-using PlatformPlatform.SharedKernel.Entities;
+using PlatformPlatform.SharedKernel.Domain;
 
 namespace PlatformPlatform.SharedKernel.DomainEvents;
